CLN: extract transform ops
