+++ tensorflow_addons/image/cc/kernels/image_ops.cc
@@ -19,11 +19,11 @@ limitations under the License.
 #define EIGEN_USE_GPU
 #endif  // GOOGLE_CUDA
 
-#include "tensorflow/contrib/image/kernels/image_ops.h"
 #include "tensorflow/core/framework/op_kernel.h"
 #include "tensorflow/core/framework/register_types.h"
 #include "tensorflow/core/framework/types.h"
 #include "tensorflow/core/platform/types.h"
+#include "tensorflow_addons/image/cc/kernels/image_ops.h"
 
 namespace tensorflow {
 
@@ -120,10 +120,6 @@ class ImageProjectiveTransform : public OpKernel {
 };
 
 #define REGISTER(TYPE)                                                \
-  REGISTER_KERNEL_BUILDER(Name("ImageProjectiveTransform")            \
-                              .Device(DEVICE_CPU)                     \
-                              .TypeConstraint<TYPE>("dtype"),         \
-                          ImageProjectiveTransform<CPUDevice, TYPE>); \
   REGISTER_KERNEL_BUILDER(Name("ImageProjectiveTransformV2")          \
                               .Device(DEVICE_CPU)                     \
                               .TypeConstraint<TYPE>("dtype"),         \
@@ -163,10 +159,6 @@ TF_CALL_double(DECLARE_FUNCTOR);
 }  // end namespace functor
 
 #define REGISTER(TYPE)                                                \
-  REGISTER_KERNEL_BUILDER(Name("ImageProjectiveTransform")            \
-                              .Device(DEVICE_GPU)                     \
-                              .TypeConstraint<TYPE>("dtype"),         \
-                          ImageProjectiveTransform<GPUDevice, TYPE>); \
   REGISTER_KERNEL_BUILDER(Name("ImageProjectiveTransformV2")          \
                               .Device(DEVICE_GPU)                     \
                               .TypeConstraint<TYPE>("dtype")          \
