+++ tensorflow_addons/image/cc/ops/image_ops.cc
@@ -1,4 +1,4 @@
-/* Copyright 2016 The TensorFlow Authors. All Rights Reserved.
+/* Copyright 2019 The TensorFlow Authors. All Rights Reserved.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -91,20 +91,6 @@ the `transforms` to the `images`. Satisfies the description above.
 
 }  // namespace
 
-// TODO(ringwalt): Add a "fill_mode" attr with "constant", "mirror", etc.
-// TODO(ringwalt): Add a "fill_constant" argument for constant mode (default 0).
-REGISTER_OP("ImageProjectiveTransform")
-    .Input("images: dtype")
-    .Input("transforms: float32")
-    .Attr("dtype: {uint8, int32, int64, float16, float32, float64}")
-    .Attr("interpolation: string")
-    .Output("transformed_images: dtype")
-    // Output shape is identical to input images.
-    .SetShapeFn([](InferenceContext* c) {
-      c->set_output(0, c->input(0));
-      return Status::OK();
-    })
-    .Doc(kImageProjectiveTransformDoc);
 
 // V2 op supports output_shape.
 REGISTER_OP("ImageProjectiveTransformV2")
@@ -116,75 +102,4 @@ REGISTER_OP("ImageProjectiveTransformV2")
     .Output("transformed_images: dtype")
     .SetShapeFn(ResizeShapeFn)
     .Doc(kImageProjectiveTransformDoc);
-
-REGISTER_OP("BipartiteMatch")
-    .Input("distance_mat: float")
-    .Input("num_valid_rows: float")
-    .Attr("top_k: int = -1")
-    .Output("row_to_col_match_indices: int32")
-    .Output("col_to_row_match_indices: int32")
-    .SetIsStateful()
-    .SetShapeFn([](InferenceContext* c) {
-      ShapeHandle input;
-      TF_RETURN_IF_ERROR(c->WithRank(c->input(0), 2, &input));
-      c->set_output(0, c->MakeShape({c->Dim(input, 0)}));
-      c->set_output(1, c->MakeShape({c->Dim(input, 1)}));
-      return Status::OK();
-    })
-    .Doc(R"doc(
-Find bipartite matching based on a given distance matrix.
-
-A greedy bi-partite matching algorithm is used to obtain the matching with the
-(greedy) minimum distance.
-
-distance_mat: A 2-D float tensor of shape `[num_rows, num_columns]`. It is a
-  pair-wise distance matrix between the entities represented by each row and
-  each column. It is an asymmetric matrix. The smaller the distance is, the more
-  similar the pairs are. The bipartite matching is to minimize the distances.
-num_valid_rows: A scalar or a 1-D tensor with one element describing the
-  number of valid rows of distance_mat to consider for the bipartite matching.
-  If set to be negative, then all rows from `distance_mat` are used.
-top_k: A scalar that specifies the number of top-k matches to retrieve.
-  If set to be negative, then is set according to the maximum number of
-  matches from `distance_mat`.
-row_to_col_match_indices: A vector of length num_rows, which is the number of
-  rows of the input `distance_matrix`.
-  If `row_to_col_match_indices[i]` is not -1, row i is matched to column
-  `row_to_col_match_indices[i]`.
-col_to_row_match_indices: A vector of length num_columns, which is the number
-  of columns of the input distance matrix.
-  If `col_to_row_match_indices[j]` is not -1, column j is matched to row
-  `col_to_row_match_indices[j]`.
-)doc");
-
-REGISTER_OP("ImageConnectedComponents")
-    .Input("image: dtype")
-    .Output("components: int64")
-    .Attr(
-        "dtype: {int64, int32, uint16, int16, uint8, int8, half, float, "
-        "double, bool, string}")
-    .SetShapeFn([](InferenceContext* c) {
-      return shape_inference::UnchangedShape(c);
-    })
-    .Doc(R"doc(
-Find the connected components of image(s).
-
-For each image (along the 0th axis), all connected components of adjacent pixels
-with the same non-zero value are detected and given unique ids.
-
-The returned `components` tensor has 0s for the zero pixels of `images`, and
-arbitrary nonzero ids for the connected components of nonzero values. Ids are
-unique across all of the images, and are in row-major order by the first pixel
-in the component.
-
-Uses union-find with union by rank but not path compression, giving a runtime of
-`O(n log n)`. See:
-    https://en.wikipedia.org/wiki/Disjoint-set_data_structure#Time_Complexity
-
-image: Image(s) with shape (N, H, W).
-components: Component ids for each pixel in "image". Same shape as "image". Zero
-    pixels all have an output of 0, and all components of adjacent pixels with
-    the same value are given consecutive ids, starting from 1.
-)doc");
-
 }  // namespace tensorflow
